MetaGPT/ephemeral_env_platform/requirements.txt
flask==1.1.2
flask-security==3.0.0
sqlalchemy==1.3.23
azure-sdk-for-python==4.0.0
docker==4.4.4
bcrypt==3.2.0